unittest methods

+++ Hello World/testing/TDD/cool_game.py
@@ -1,4 +1,9 @@
+from random import choice
+
+
 def greet(name, isEnemy):
+    if not isinstance(isEnemy, bool):
+        raise ValueError('isEnemy must be a boolean type')
     if isEnemy:
         return f'Hello {name}! I will kill bastard!'
     else:
@@ -10,3 +15,14 @@ def eat_burgers(number):
         return 'Oh! I overate!'
     else:
         return 'Mmm! Thas was excellent!'
+    
+    
+def can_fly(name):
+    if name == 'Batman':
+        return True
+    else:
+        return False
+    
+    
+def get_arsenal():
+    return choice(('knife', 'handgun', 'machine gun'))
